Implement custom dropdown hide behavior
The autoHide feature of the PopUp widget is insufficient as it closes in
cases it should not, e.g. clicking on the scrollbar or textfield, but
not on pressing Enter or Escape. The custom implementation is based on
focus events and hides if neither the dropdown, nor the linked control
or any widget within the linked control is focused. Automatically hiding
the dropdown on Enter or Escape is difficult since it may not always be
focused, and may also not always desired. It's currently achieved
partially by clientscripting using key events and DefaultSelection
events. DropDown does not (yet) support key events though, and can
therefore not be closed if ESC is pressed while it's focused unless
either key events or auto-hide on ESC are implemented.