Rename ModelListener.js to AutoSuggest.js
ModelListner.js could be confused as the counterpart to
ModelListener.java, which it is not. AutoSuggest.js may not be ideal
either, but it is roughly the counterpart to AutoSuggest.java. No
"Listener" postfix was used since it is not a listener on AutoSuggest,
but defines the AutoSugges behavior itself.