Fix AutoSuggest.setDataProvider
Because the dataProvider id is set before either listeners not the
dataProvider itself are created the change:dataProvider listener isn't
called, and even if it would it could not get the dataProvider. This
only works if the dataProvider is changed later on. So in addition, the
dataProvider is looked up before the filter processes a user input.