Create protected methods in AutoSuggest to allow custom scripts
"attachClientListeners" gets model, text and dropDown as arguments to
allow attaching listeners. The ClientListeners attached to Text should
be given as a return value because they need to be removed from Text on
dispose. (The other listeners are removed implicitly by DropDown and
Model.)

"removeTextClientListeners" is called on dispose with the value returned
from "attachClientListeners" and the Text widget.