Force DropDown to appear and position after parents have been layouted
Previously the DropDown could be positioned incorrectly if the parent
changed it's absolute position without changing it's parent-relative
position. The only known way to compensate for this is to re-position
it in the appear event. However, this event could be fired before the
parent has been layouted if visibility is used to hide the popup, so we
will now use display.