Update tests and demo to Java 8
Since we the testfixture is now compiled to Java 8, the tests on
incubator projects need to be updated as well.

As we did in RAP itself, update all bundles that don't go into the
runtime to Java 8 in order to avoid problems with dependencies to
other bundles that require Java 8.